+++ crates/rust-client/src/sync/block_headers.rs
@@ -46,12 +46,11 @@ impl<R: FeltRng> Client<R> {
     /// Attempts to retrieve the genesis block from the store. If not found,
     /// it requests it from the node and store it.
     pub(crate) async fn ensure_genesis_in_place(&mut self) -> Result<BlockHeader, ClientError> {
-        let genesis = self.store.get_block_header_by_num(0).await;
+        let genesis = self.store.get_block_header_by_num(0).await?;
 
         match genesis {
-            Ok((block, _)) => Ok(block),
-            Err(StoreError::BlockHeaderNotFound(0)) => self.retrieve_and_store_genesis().await,
-            Err(err) => Err(ClientError::StoreError(err)),
+            Some((block, _)) => Ok(block),
+            None => self.retrieve_and_store_genesis().await,
         }
     }
 
@@ -116,11 +115,10 @@ impl<R: FeltRng> Client<R> {
         let current_peaks = self.store.get_chain_mmr_peaks_by_block_num(current_block_num).await?;
 
         let track_latest = if current_block_num != 0 {
-            match self.store.get_block_header_by_num(current_block_num - 1).await {
-                Ok((_, previous_block_had_notes)) => Ok(previous_block_had_notes),
-                Err(StoreError::BlockHeaderNotFound(_)) => Ok(false),
-                Err(err) => Err(ClientError::StoreError(err)),
-            }?
+            match self.store.get_block_header_by_num(current_block_num - 1).await? {
+                Some((_, previous_block_had_notes)) => previous_block_had_notes,
+                None => false,
+            }
         } else {
             false
         };
@@ -129,8 +127,11 @@ impl<R: FeltRng> Client<R> {
             PartialMmr::from_parts(current_peaks, tracked_nodes, track_latest);
 
         if include_current_block {
-            let (current_block, has_client_notes) =
-                self.store.get_block_header_by_num(current_block_num).await?;
+            let (current_block, has_client_notes) = self
+                .store
+                .get_block_header_by_num(current_block_num)
+                .await?
+                .expect("Current block should be in the store");
 
             current_partial_mmr.add(current_block.hash(), has_client_notes);
         }
@@ -149,7 +150,11 @@ impl<R: FeltRng> Client<R> {
     ) -> Result<BlockHeader, ClientError> {
         if current_partial_mmr.is_tracked(block_num as usize) {
             warn!("Current partial MMR already contains the requested data");
-            let (block_header, _) = self.store.get_block_header_by_num(block_num).await?;
+            let (block_header, _) = self
+                .store
+                .get_block_header_by_num(block_num)
+                .await?
+                .expect("Block header should be tracked");
             return Ok(block_header);
         }
         let (block_header, mmr_proof) = self.rpc_api.get_block_header_with_proof(block_num).await?;
@@ -184,7 +189,9 @@ impl<R: FeltRng> Client<R> {
         let epoch = block_epoch_from_number(block_num);
         let epoch_block_number = block_num_from_epoch(epoch);
 
-        if let Ok((epoch_block, _)) = self.store.get_block_header_by_num(epoch_block_number).await {
+        if let Some((epoch_block, _)) =
+            self.store.get_block_header_by_num(epoch_block_number).await?
+        {
             return Ok(epoch_block);
         }