+++ crates/rust-client/src/tests.rs
@@ -151,7 +151,7 @@ async fn test_get_input_note() {
         .unwrap();
 
     // retrieve note from database
-    let retrieved_note = client.get_input_note(original_note.id()).await.unwrap();
+    let retrieved_note = client.get_input_note(original_note.id()).await.unwrap().unwrap();
 
     let recorded_note: InputNoteRecord = original_note.into();
     assert_eq!(recorded_note.id(), retrieved_note.id());
@@ -172,7 +172,7 @@ async fn insert_basic_account() {
     let fetched_account_data = client.get_account(account.id()).await;
     assert!(fetched_account_data.is_ok());
 
-    let fetched_account = fetched_account_data.unwrap();
+    let fetched_account = fetched_account_data.unwrap().unwrap();
     let fetched_account_seed = fetched_account.seed().cloned();
     let fetched_account: Account = fetched_account.into();
 
@@ -203,7 +203,7 @@ async fn insert_faucet_account() {
     let fetched_account_data = client.get_account(account.id()).await;
     assert!(fetched_account_data.is_ok());
 
-    let fetched_account = fetched_account_data.unwrap();
+    let fetched_account = fetched_account_data.unwrap().unwrap();
     let fetched_account_seed = fetched_account.seed().cloned();
     let fetched_account: Account = fetched_account.into();
 
@@ -270,7 +270,7 @@ async fn test_account_code() {
         .add_account(&account, Some(Word::default()), &AuthSecretKey::RpoFalcon512(key_pair), false)
         .await
         .unwrap();
-    let retrieved_acc = client.get_account(account.id()).await.unwrap();
+    let retrieved_acc = client.get_account(account.id()).await.unwrap().unwrap();
     assert_eq!(*account.code(), *retrieved_acc.account().code());
 }
 
@@ -294,14 +294,14 @@ async fn test_get_account_by_id() {
 
     // Retrieving an existing account should succeed
     let (acc_from_db, _account_seed) = match client.get_account_header_by_id(account.id()).await {
-        Ok(account) => account,
+        Ok(account) => account.unwrap(),
         Err(err) => panic!("Error retrieving account: {}", err),
     };
     assert_eq!(AccountHeader::from(account), acc_from_db);
 
     // Retrieving a non existing account should fail
     let invalid_id = AccountId::try_from(ACCOUNT_ID_FUNGIBLE_FAUCET_ON_CHAIN_2).unwrap();
-    assert!(client.get_account_header_by_id(invalid_id).await.is_err());
+    assert!(client.get_account_header_by_id(invalid_id).await.unwrap().is_none());
 }
 
 #[tokio::test]
@@ -579,7 +579,7 @@ async fn test_import_processing_note_returns_error() {
     client.submit_transaction(transaction).await.unwrap();
 
     let note_id = transaction_request.expected_output_notes().next().unwrap().id();
-    let note = client.get_input_note(note_id).await.unwrap();
+    let note = client.get_input_note(note_id).await.unwrap().unwrap();
 
     let input = [(note.try_into().unwrap(), None)];
     let consume_note_request =